#3 performed data cleaning/processing on the message data
